# rpc : add support for multiple devices