--- a/common/arg.cpp
+++ b/common/arg.cpp
@@ -1721,17 +1721,26 @@ static void add_rpc_devices(const std::string & servers) {
     if (!rpc_reg) {
         throw std::invalid_argument("failed to find RPC backend");
     }
-    typedef ggml_backend_dev_t (*ggml_backend_rpc_add_device_t)(const char * endpoint);
+    typedef int (*ggml_backend_rpc_get_device_count_t)(const char * endpoint);
+    ggml_backend_rpc_get_device_count_t ggml_backend_rpc_get_device_count_fn =
+        (ggml_backend_rpc_get_device_count_t) ggml_backend_reg_get_proc_address(rpc_reg, "ggml_backend_rpc_get_device_count");
+    if (!ggml_backend_rpc_get_device_count_fn) {
+        throw std::invalid_argument("failed to find RPC device count function");
+    }
+    typedef ggml_backend_dev_t (*ggml_backend_rpc_add_device_t)(const char * endpoint, uint32_t device);
     ggml_backend_rpc_add_device_t ggml_backend_rpc_add_device_fn = (ggml_backend_rpc_add_device_t) ggml_backend_reg_get_proc_address(rpc_reg, "ggml_backend_rpc_add_device");
     if (!ggml_backend_rpc_add_device_fn) {
         throw std::invalid_argument("failed to find RPC device add function");
     }
     for (const auto & server : rpc_servers) {
-        ggml_backend_dev_t dev = ggml_backend_rpc_add_device_fn(server.c_str());
-        if (dev) {
-            ggml_backend_device_register(dev);
-        } else {
-            throw std::invalid_argument("failed to register RPC device");
+        int dev_count = ggml_backend_rpc_get_device_count_fn(server.c_str());
+        for (int i = 0; i < dev_count; i++) {
+            ggml_backend_dev_t dev = ggml_backend_rpc_add_device_fn(server.c_str(), i);
+            if (dev) {
+                ggml_backend_device_register(dev);
+            } else {
+                throw std::invalid_argument("failed to register RPC device");
+            }
         }
     }
 }