--- a/common/arg.cpp
+++ b/common/arg.cpp
@@ -1721,27 +1721,14 @@ static void add_rpc_devices(const std::string & servers) {
     if (!rpc_reg) {
         throw std::invalid_argument("failed to find RPC backend");
     }
-    typedef int (*ggml_backend_rpc_get_device_count_t)(const char * endpoint);
-    ggml_backend_rpc_get_device_count_t ggml_backend_rpc_get_device_count_fn =
-        (ggml_backend_rpc_get_device_count_t) ggml_backend_reg_get_proc_address(rpc_reg, "ggml_backend_rpc_get_device_count");
-    if (!ggml_backend_rpc_get_device_count_fn) {
-        throw std::invalid_argument("failed to find RPC device count function");
-    }
-    typedef ggml_backend_dev_t (*ggml_backend_rpc_add_device_t)(const char * endpoint, uint32_t device);
-    ggml_backend_rpc_add_device_t ggml_backend_rpc_add_device_fn = (ggml_backend_rpc_add_device_t) ggml_backend_reg_get_proc_address(rpc_reg, "ggml_backend_rpc_add_device");
-    if (!ggml_backend_rpc_add_device_fn) {
-        throw std::invalid_argument("failed to find RPC device add function");
+    typedef ggml_backend_reg_t (*ggml_backend_rpc_add_server_t)(const char * endpoint);
+    ggml_backend_rpc_add_server_t ggml_backend_rpc_add_server_fn = (ggml_backend_rpc_add_server_t) ggml_backend_reg_get_proc_address(rpc_reg, "ggml_backend_rpc_add_server");
+    if (!ggml_backend_rpc_add_server_fn) {
+        throw std::invalid_argument("failed to find RPC add server function");
     }
     for (const auto & server : rpc_servers) {
-        int dev_count = ggml_backend_rpc_get_device_count_fn(server.c_str());
-        for (int i = 0; i < dev_count; i++) {
-            ggml_backend_dev_t dev = ggml_backend_rpc_add_device_fn(server.c_str(), i);
-            if (dev) {
-                ggml_backend_device_register(dev);
-            } else {
-                throw std::invalid_argument("failed to register RPC device");
-            }
-        }
+        auto reg = ggml_backend_rpc_add_server_fn(server.c_str());
+        ggml_backend_register(reg);
     }
 }