@rgerganov 2025-09-30
I am also considering changing the naming scheme for RPC devices. Now I am using `RPC<X>[<host>:<port>]` for both `name` and `description` which doesn't look nice in various logs.

I am thinking of switching to `RPC<X>` for device name and `[<host>:<port>]` for device description. The `X` number should be global monotonic counter making RPC devices appear like this:
```
RPC0 ([localhost:50052])
RPC1 ([localhost:50052])
RPC2 ([localhost:18053])
RPC3 ([localhost:18054])
...
```

Thoughts?