@rgerganov 2025-10-04
I did some benchmarks with two hosts connected on a local gigabit network. The first host has no accelerators and offloads everything to the second host which has two NVIDIA cards (RTX 5090, GTX 1660).

## master

```
$ ./llama-bench -m ~/llamacpp/models/gemma-3-4b-it-q4_0.gguf -ngl 99 --rpc 192.168.88.34:50052,192.168.88.34:50053 --list-devices
load_backend: loaded RPC backend from /home/deck/bench/master/libggml-rpc.so
load_backend: loaded CPU backend from /home/deck/bench/master/libggml-cpu-haswell.so
Available devices:
  RPC[192.168.88.34:50052]: RPC[192.168.88.34:50052] (32109 MiB, 31588 MiB free)
  RPC[192.168.88.34:50053]: RPC[192.168.88.34:50053] (5748 MiB, 5603 MiB free)
```

| model                          |       size |     params | backend    | ngl |            test |                  t/s |
| ------------------------------ | ---------: | ---------: | ---------- | --: | --------------: | -------------------: |
| gemma3 4B Q4_0                 |   2.93 GiB |     3.88 B | RPC        |  99 |           pp512 |       1039.83 ± 0.44 |
| gemma3 4B Q4_0                 |   2.93 GiB |     3.88 B | RPC        |  99 |           tg128 |         33.18 ± 0.05 |

## PR

```
$ ./llama-bench -m ~/llamacpp/models/gemma-3-4b-it-q4_0.gguf -ngl 99 --rpc 192.168.88.34:50052 --list-devices
load_backend: loaded RPC backend from /home/deck/bench/multidev/libggml-rpc.so
load_backend: loaded CPU backend from /home/deck/bench/multidev/libggml-cpu-haswell.so
Available devices:
  RPC0: 192.168.88.34:50052 (32109 MiB, 31588 MiB free)
  RPC1: 192.168.88.34:50052 (5748 MiB, 5671 MiB free)
```

| model                          |       size |     params | backend    | ngl |            test |                  t/s |
| ------------------------------ | ---------: | ---------: | ---------- | --: | --------------: | -------------------: |
| gemma3 4B Q4_0                 |   2.93 GiB |     3.88 B | RPC        |  99 |           pp512 |       1266.96 ± 2.77 |
| gemma3 4B Q4_0                 |   2.93 GiB |     3.88 B | RPC        |  99 |           tg128 |         33.68 ± 0.10 |


# Exposing the same card (RTX 5090) twice

## master

```
$ ./llama-bench -m ~/llamacpp/models/gemma-3-4b-it-q4_0.gguf -ngl 99 --rpc 192.168.88.34:50052,192.168.88.34:50053 --list-devices
load_backend: loaded RPC backend from /home/deck/bench/master/libggml-rpc.so
load_backend: loaded CPU backend from /home/deck/bench/master/libggml-cpu-haswell.so
Available devices:
  RPC[192.168.88.34:50052]: RPC[192.168.88.34:50052] (32109 MiB, 31588 MiB free)
  RPC[192.168.88.34:50053]: RPC[192.168.88.34:50053] (32109 MiB, 31084 MiB free)
```

| model                          |       size |     params | backend    | ngl |            test |                  t/s |
| ------------------------------ | ---------: | ---------: | ---------- | --: | --------------: | -------------------: |
| gemma3 4B Q4_0                 |   2.93 GiB |     3.88 B | RPC        |  99 |           pp512 |       1424.32 ± 2.49 |
| gemma3 4B Q4_0                 |   2.93 GiB |     3.88 B | RPC        |  99 |           tg128 |         47.10 ± 0.02 |

## PR

```
$ bin/rpc-server -c -d CUDA0,CUDA0 -H 0.0.0.0 -p 50052
```

```
$ ./llama-bench -m ~/llamacpp/models/gemma-3-4b-it-q4_0.gguf -ngl 99 --rpc 192.168.88.34:50052 --list-devices
load_backend: loaded RPC backend from /home/deck/bench/multidev/libggml-rpc.so
load_backend: loaded CPU backend from /home/deck/bench/multidev/libggml-cpu-haswell.so
Available devices:
  RPC0: 192.168.88.34:50052 (32109 MiB, 31588 MiB free)
  RPC1: 192.168.88.34:50052 (32109 MiB, 31588 MiB free)
```

| model                          |       size |     params | backend    | ngl |            test |                  t/s |
| ------------------------------ | ---------: | ---------: | ---------- | --: | --------------: | -------------------: |
| gemma3 4B Q4_0                 |   2.93 GiB |     3.88 B | RPC        |  99 |           pp512 |       1896.91 ± 5.12 |
| gemma3 4B Q4_0                 |   2.93 GiB |     3.88 B | RPC        |  99 |           tg128 |         47.65 ± 0.49 |


In summary, this patch brings massive improvements in PP and some small improvements in TG when multiple devices are used with RPC.
I also don't see any regressions when using RPC with a single device.

However, I noticed that I made the mistake of putting `RPC_CMD_DEVICE_COUNT` first and thus changing the number for `RPC_CMD_HELLO` resulting in bad user experience when the client and server version mismatch. I will fix this and then merge.