@rgerganov 2025-10-04
The old behavior was to copy tensors from device 0 to main host and the from main host to device 1 and so on. The new behavior is that main host sends `RPC_CMD_COPY_TENSOR` to copy tensor from dev 0 to dev 1 which saves a lot of network traffic (tensor data is not sent). You can see the commands being received and executed by setting `GGML_RPC_DEBUG=1` before starting `rpc-server`. I will also update the documentation soon.

> I'm away from home, but can test this exact setup when I get back with `deepseek-v3` and will post the results.

That would be very much appreciated, thanks