@slaren 2025-10-04
There is a possible issue here:
```C++
static bool ggml_backend_rpc_buffer_cpy_tensor(ggml_backend_buffer_t buffer, const ggml_tensor * src, ggml_tensor * dst) {
    // check if src and dst are on the same server
    ggml_backend_buffer_t src_buffer = src->buffer;
    ggml_backend_rpc_buffer_context * src_ctx = (ggml_backend_rpc_buffer_context *)src_buffer->context;
    ggml_backend_buffer_t dst_buffer = dst->buffer;
    ggml_backend_rpc_buffer_context * dst_ctx = (ggml_backend_rpc_buffer_context *)dst_buffer->context;
    if (src_ctx->sock != dst_ctx->sock) {
        return false;
    }
```
Only the `dst` tensor is guaranteed to be on a RPC buffer, the `src` tensor could be on any other buffer type. Before casting the `src` context to `ggml_backend_rpc_buffer_context`, it is necessary to check that it is an RPC buffer.