@rgerganov 2025-10-04
> Only the `dst` tensor is guaranteed to be on a RPC buffer, the `src` tensor could be on any other buffer type. Before casting the `src` context to `ggml_backend_rpc_buffer_context`, it is necessary to check that it is an RPC buffer.

Submitted #16421