@slaren 2025-10-08
I don't think the comparison makes much sense. If we want to improve the efficiency of the RPC backend, we need to look at what operations introduce latency. I suspect that one of the biggest sources of latency is that every command blocks until a response is received from the server, but this may not be strictly necessary in every case. For instance, async operations should not require waiting for a response from the server in most cases. Implementing async support, while also moving the network to a different thread (such that async operations return immediately), could reduce the number of roundtrips and reduce latency.